#125 | Flux refactor
