+++ application/flux/constants.js
