+++ application/flux/stores.js
@@ -1,0 +1,9 @@
+'use strict';
+
+let TokenStore = require('./user/token-store');
+let UserStore  = require('./user/user-store');
+
+module.exports = {
+    Token : TokenStore,
+    User  : UserStore
+};
