+++ application/flux/user/oauth-client.js
@@ -1,6 +1,6 @@
 'use strict';
 
-let config      = require('../config');
+let config      = require('../../config');
 let HttpGateway = require('synapse-common/http/auth-gateway');
 let qs          = require('querystring');
 let store       = require('store');
