--- application/stores.js
@@ -1,9 +1,0 @@
-'use strict';
-
-let TokenStore = require('./store/token');
-let UserStore  = require('./store/user');
-
-module.exports = {
-    TokenStore : TokenStore,
-    UserStore  : UserStore
-};
